Puts command classes in separate headers from their respective abstractions.
Names all command headers with _cmd and adds include guards.

git-pr-chain: debug-cmd-headers
