+++ controller/lib/debug/peek_cmd.cpp
@@ -13,16 +13,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-/*
- * This file implements a 'poke' command which allows us
- * to write to wrap values in memory.  Use with caution!
- *
- * The data passed to the command consists of a 32-bit
- * starting address and one or more data bytes to be
- * written to consecutive addresses starting at the
- * given address
- */
+#include "peek_cmd.h"
 
-#include "poke.h"
-
-PokeCmd pokeCmd;
+PeekCmd peekCmd;
